pyproject.toml
[project]
name = "hf-agents-assignment"
version = "0.1.0"
description = ""
authors = [
    {name = "AnnaErsh",email = "aershova096@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "llama-index (>=0.12.34,<0.13.0)",
    "gradio (>=5.29.0,<6.0.0)",
    "llama-index-vector-stores-chroma (>=0.4.1,<0.5.0)",
    "llama-index-utils-workflow (>=0.3.1,<0.4.0)",
    "llama-index-llms-huggingface-api (>=0.4.2,<0.5.0)",
    "duckduckgo-search (>=6.1.0,<7.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "yt-dlp (>=2025.4.30,<2026.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "openai-whisper (>=20240930,<20240931)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "llama-index-tools-duckduckgo (>=0.3.0,<0.4.0)",
    "llama-index-tools-wikipedia (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
